requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-cors-headers==3.8.0
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
flake8==3.9.2
idna==2.10
iniconfig==1.1.1
Markdown==3.3.4
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
